Fix Qt5 and buildroot cross-compilation compatibility issues
This commit fixes several issues that prevented successful cross-compilation
with buildroot for Raspberry Pi 5:

file_manager module:
- Add Qt5::Concurrent dependency to CMakeLists.txt
- Add missing QDateTime and QVariant includes
- Delete move constructors/assignment operators from QObject-derived classes
  (recording_scanner, FileBrowserController) - QObject is non-copyable/movable
- Add missing private method declarations (applyFilters, updateStorageInfo)
- Fix Qt6-only toInteger() calls to Qt5-compatible toVariant().toLongLong()
- Add missing <set>, <sstream>, <iomanip> includes
- Fix -Wdouble-promotion warning in metadata_extractor

network_manager module:
- Fix middleware_functors signature to match piejam redux pattern
- Fix aggregate initialization warnings for nfs_mount_state struct
- Fix -Wignored-attributes warning with unique_ptr pclose deleter
- Fix wifi_connection_result usage (struct, not enum)

These fixes ensure the code compiles with GCC's strict warnings enabled
(-Wall -Wextra -Werror) in the buildroot cross-compilation environment.

+++ piejam_libs/file_manager/src/piejam/file_manager/metadata_extractor.cpp
@@ -7,6 +7,7 @@
 #include <sndfile.h>
 
 #include <QCryptographicHash>
+#include <QDateTime>
 #include <QFile>
 #include <QFileInfo>
 
@@ -225,7 +226,7 @@ metadata_extractor::analyze_levels(std::string const& file_path,
                 peak_level = sample;
             }
 
-            sum_squares += static_cast<double>(sample) * sample;
+            sum_squares += static_cast<double>(sample) * static_cast<double>(sample);
 
             if (sample >= clip_threshold)
             {
